requirements.txt
python-telegram-bot==20.7
pysnmp==4.4.12